Inital commit
This is the code I've been using successfully for a while.
It is untested and not very well structured, but it works.
It must be cloned into your `~/.emacs.d/elpa/` directory as
`keepass-20190830.0001` and then, you'll need to include the following
lines in your `~/.emacs.d/init.el`:

```emacs-lisp
(require 'keepass-mode)
```