Use `shell-quote-argument` to protect command from failing
This makes sense only for the `keepass-mode-password` argument, since
it is provided manually by the user and it's probable to contain
special characters.
The argument `keepass-mode-db` is a file name provided by the user
using the minibuffer. Also quoting this argument prevents the command
to work well.
The arguments `command` and `group` are provided by the code, not the
user. Also, the `group` argument is quoted conveniently.