Fix tests on updated keepassxc-cli
Latest keepassxc-cli hides passwords in output unless the `-s` option
is used with `show` command. This was fixed in a previous commit, but
the tests are not adapted to the new keepassxc-cli behavior.

This latest update also changed how the password is requested, using a
different sentence and the stderr stream.